# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"Famigo API","description":"家族向けお出かけスポット共有サービス Famigo（ファミゴー）のAPI仕様書です。"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"レビュー","description":"スポットのレビュー一覧 / 投稿 / 編集 / 削除"},{"name":"カテゴリ","description":"カテゴリ一覧"},{"name":"外部認証","description":"ユーザー登録 / ログイン（JWT発行）"},{"name":"お気に入り","description":"お気に入り一覧 / 追加 / 削除"},{"name":"スポット","description":"スポット一覧 / 詳細"},{"name":"ユーザー情報管理","description":"ログイン中ユーザーの取得 / 更新 / 退会"}],"paths":{"/api/users/me":{"get":{"tags":["ユーザー情報管理"],"summary":"自分のユーザー情報を取得（Read）","description":"ログイン中ユーザーの基本情報（id / name / email / role）を返します。","operationId":"me","responses":{"200":{"description":"取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"put":{"tags":["ユーザー情報管理"],"summary":"自分のプロフィールを変更（Update）（表示名+メール）","description":"ログイン中ユーザーの表示名（users.name）とメールアドレス（users.email）を同時に更新します。 方針（原子性）：入力エラーがある場合は、どの項目も更新しません。","operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserMeRequest"}}},"required":true},"responses":{"200":{"description":"更新成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeResponse"}}}},"400":{"description":"入力不正（バリデーションエラー）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"更新失敗（email重複）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["ユーザー情報管理"],"summary":"退会（論理削除）（Delete）","description":"ログイン中ユーザーを退会（users.is_deleted=1）にし、表示名を「退会ユーザー」に更新します。レビューは残ります。","operationId":"withdraw","responses":{"204":{"description":"退会成功（No Content）"},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/users/me/password":{"put":{"tags":["ユーザー情報管理"],"summary":"自分のパスワードを変更（Update）","description":"ログイン中ユーザーのパスワード（users.password_hash）を更新します。","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserMeRequest"}}},"required":true},"responses":{"204":{"description":"更新成功（No Content）"},"400":{"description":"入力不正（バリデーションエラー）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/spots/{spotId}/reviews/{reviewId}":{"put":{"tags":["レビュー"],"summary":"レビュー編集【スポットID + レビューID指定】","description":"パスで指定されたスポットID配下のレビューを編集します。USER は自分のレビューのみ、ADMIN は全レビューを編集できます。","operationId":"updateReview","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"reviewId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewUpsertRequest"}}},"required":true},"responses":{"204":{"description":"レビュー編集成功（No Content）"},"400":{"description":"入力不正（バリデーションエラー）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（本人以外など）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"スポット or レビューが存在しない","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["レビュー"],"summary":"レビュー削除【スポットID + レビューID指定】","description":"パスで指定されたスポットID配下のレビューを削除します（論理削除）。USER は自分のレビューのみ、ADMIN は全レビューを削除できます。","operationId":"deleteReview","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"reviewId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"レビュー削除成功（No Content）"},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（本人以外など）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"404":{"description":"スポット or レビューが存在しない","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/users":{"post":{"tags":["外部認証"],"summary":"ユーザー登録（サインアップ）","description":"displayName / email / password / passwordConfirm を受け取り、ユーザーを作成します。登録成功時はJWTを発行して返却します。","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"登録成功（JWT発行）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"400":{"description":"入力不正（バリデーションエラー）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"409":{"description":"登録失敗（email重複）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/spots/{spotId}/reviews":{"get":{"tags":["レビュー"],"summary":"レビュー一覧取得【スポットID指定】","description":"パスで指定されたスポットIDに紐づくレビュー一覧を取得します。","operationId":"getReviewsBySpotId","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"レビュー一覧取得成功","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReviewListItemDto"}}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"post":{"tags":["レビュー"],"summary":"スポット1件についてレビュー投稿【スポットID指定】","description":"パスで指定されたスポットIDに対して、ログイン中ユーザーがレビューを投稿します。","operationId":"createReview","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewUpsertRequest"}}},"required":true},"responses":{"204":{"description":"レビュー投稿成功（No Content）"},"400":{"description":"入力不正（バリデーションエラー）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/favorites/{spotId}":{"post":{"tags":["お気に入り"],"summary":"スポット1件をお気に入り登録（Create/復活）","description":"パスで指定されたスポットIDをお気に入り登録します。既に存在し論理削除されている場合は復活します。","operationId":"addFavorite","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"お気に入り登録成功（No Content）"},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}},"delete":{"tags":["お気に入り"],"summary":"スポット1件をお気に入り解除（論理削除）（Delete）","description":"パスで指定されたスポットIDをお気に入り解除（論理削除）します。","operationId":"removeFavorite","parameters":[{"name":"spotId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"お気に入り解除成功（No Content）"},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/auth/token":{"post":{"tags":["外部認証"],"summary":"ログイン（JWT発行）","description":"email / password が正しい場合に JWT を発行します。","operationId":"token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"ログイン成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"400":{"description":"入力不正（バリデーションエラー）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"認証失敗（email/password不一致）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/spots":{"get":{"tags":["スポット"],"summary":"スポット一覧の取得【検索条件指定可】","description":"カテゴリ情報・設備情報を JOIN 済みのスポット一覧を取得します。クエリパラメータでカテゴリID、キーワード、予算（Enum名）、対象年齢（Enum名）、設備などを指定可能です。いずれの条件も指定されない場合は、論理削除されていないスポットを ID 昇順で全件返します。ログイン時は、お気に入り状態（isFavorite）も合わせて返します。","operationId":"getSpots","parameters":[{"name":"categoryIds","in":"query","required":false,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"price","in":"query","required":false,"schema":{"type":"array","items":{"type":"string","description":"料金区分（スポット検索条件や表示に利用する）","enum":["FREE","UNDER_1000","UNDER_2000","OVER_2000"]}}},{"name":"age","in":"query","required":false,"schema":{"type":"array","items":{"type":"string","description":"対象年齢（スポット検索条件や表示に利用する）","enum":["ALL","PRESCHOOL","ELE_LOW","ELE_HIGH","JUNIOR_HIGH"]}}},{"name":"facilities","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"一覧取得成功","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpotListItemDto"}}}}},"400":{"description":"入力不正（クエリパラメータのバリデーションエラー等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/spots/{id}":{"get":{"tags":["スポット"],"summary":"スポット詳細情報の取得【スポットID指定】","description":"パスで指定されたスポットIDに該当するスポット詳細情報を1件取得します。ログイン時は、お気に入り状態（isFavorite）も合わせて返します。","operationId":"getSpotDetail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"詳細取得成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpotDetailDto"}}}},"404":{"description":"スポットが存在しない","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/favorites":{"get":{"tags":["お気に入り"],"summary":"お気に入り一覧を取得（Read）","description":"ログイン中ユーザーのお気に入り（favorites.is_deleted=0）かつスポット有効（spots.is_deleted=0）の一覧を返します。 並び順は favorites.updated_at DESC（なければ created_at DESC）です。","operationId":"getFavorites","responses":{"200":{"description":"お気に入り一覧取得成功","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpotListItemDto"}}}}},"401":{"description":"未ログイン / トークン不正","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"権限不足（GUEST等）","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/api/categories":{"get":{"tags":["カテゴリ"],"summary":"カテゴリ一覧の取得","description":"categories テーブルのカテゴリ一覧を ID 昇順で取得します。","operationId":"getCategories","responses":{"200":{"description":"一覧取得成功","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryDto"}}}}},"500":{"description":"想定外エラー","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}}},"components":{"schemas":{"MeResponse":{"type":"object","description":"ログイン中ユーザー情報DTO","properties":{"id":{"type":"integer","format":"int64","description":"ユーザーID","example":1},"name":{"type":"string","description":"ユーザー名（ニックネーム可）","example":"デモユーザー"},"email":{"type":"string","description":"メールアドレス（ログインID）","example":"demo@example.com"},"role":{"type":"string","description":"ユーザーロール（GUEST / USER / ADMIN）","example":"USER"}}},"ErrorResponse":{"type":"object","properties":{"errorCode":{"type":"string"},"message":{"type":"string"},"fieldErrors":{"type":"array","items":{"$ref":"#/components/schemas/FieldErrorItem"}}}},"FieldErrorItem":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"UpdateUserMeRequest":{"type":"object","description":"ユーザー自己情報更新リクエスト（用途により必須項目が変わる）","properties":{"displayName":{"type":"string","description":"新しい表示名（ニックネーム可）","example":"ゆうパパ"},"email":{"type":"string","description":"新しいメールアドレス（ログインID）","example":"newmail@example.com"},"currentPassword":{"type":"string","description":"現在のパスワード","example":"Passw0rd!"},"newPassword":{"type":"string","description":"新しいパスワード（6文字以上）","example":"NewPassw0rd!"},"newPasswordConfirm":{"type":"string","description":"新しいパスワード確認","example":"NewPassw0rd!"},"newPasswordMatched":{"type":"boolean"}}},"ReviewUpsertRequest":{"type":"object","description":"レビュー投稿（Create）用リクエストデータモデル","properties":{"childAgeGroup":{"type":"string","description":"子どもの年齢帯（単一選択）","enum":["PRESCHOOL","ELE_LOW","ELE_HIGH","JUNIOR_HIGH_PLUS"],"example":"PRESCHOOL"},"rating":{"type":"integer","format":"int32","description":"総合評価（1〜5）","example":5,"maximum":5,"minimum":1},"ratingCost":{"type":"integer","format":"int32","description":"コスパ評価（1〜5）","example":4,"maximum":5,"minimum":1},"crowdLevel":{"type":"integer","format":"int32","description":"混雑度（1〜5）","example":3,"maximum":5,"minimum":1},"toiletCleanliness":{"type":"integer","format":"int32","description":"トイレ清潔度（1〜5）","example":4,"maximum":5,"minimum":1},"strollerEase":{"type":"integer","format":"int32","description":"ベビーカーの使いやすさ（1〜5）","example":5,"maximum":5,"minimum":1},"reviewText":{"type":"string","description":"レビュー本文","example":"子どもが1日中遊べて大満足でした","minLength":1},"costTotal":{"type":"integer","format":"int32","description":"かかった合計金額（円）","example":1000,"minimum":0}},"required":["childAgeGroup","rating","reviewText"]},"LoginResponse":{"type":"object","description":"ログインレスポンスDTO（JWT発行結果）","properties":{"accessToken":{"type":"string","description":"JWTアクセストークン","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"tokenType":{"type":"string","description":"アクセストークンの種別","example":"Bearer"},"expiresIn":{"type":"integer","format":"int64","description":"アクセストークンの有効期限（秒）。発行時点からの秒数","example":3600},"user":{"$ref":"#/components/schemas/MeResponse","description":"ログインユーザー情報（項目はMeResponseのDTOで別途定義）"}}},"RegisterRequest":{"type":"object","description":"ユーザー登録（サインアップ）要求","properties":{"displayName":{"type":"string","description":"表示名（ニックネーム可）","example":"ゆうパパ"},"email":{"type":"string","description":"メールアドレス（ログインID）","example":"test@example.com"},"password":{"type":"string","description":"パスワード（6文字以上）","example":"Passw0rd!"},"passwordConfirm":{"type":"string","description":"パスワード確認","example":"Passw0rd!"},"passwordMatched":{"type":"boolean"}}},"LoginRequest":{"type":"object","description":"ログインリクエストDTO","properties":{"email":{"type":"string","format":"email","description":"メールアドレス（ログインID）","example":"demo@example.com","minLength":1},"password":{"type":"string","description":"パスワード（平文）","example":"demo1234","minLength":1}},"required":["email","password"]},"SpotListItemDto":{"type":"object","description":"スポット一覧カード1件分のデータモデル","properties":{"id":{"type":"integer","format":"int64","description":"スポットID（spots.id）","example":2},"name":{"type":"string","description":"スポット名（spots.name）","example":"滋賀県希望ヶ丘文化公園"},"address":{"type":"string","description":"住所（spots.address）","example":"滋賀県野洲市北桜978"},"area":{"type":"string","description":"エリア名（spots.area）","example":"野洲市"},"priceType":{"type":"string","description":"予算区分（spots.price_type）","example":"1000円以内"},"categoryName":{"type":"string","description":"カテゴリ名（categories.name）","example":"公園"},"targetAge":{"type":"string","description":"対象年齢（spots.target_age）","example":"小学校低学年まで"},"googleMapUrl":{"type":"string","description":"GoogleマップURL（spots.google_map_url）"},"isFavorite":{"type":"boolean","description":"お気に入り済みかどうか（favorites）","example":true},"diaperChanging":{"type":"boolean","description":"オムツ替えスペースの有無（diaper_changing）","example":true},"strollerOk":{"type":"boolean","description":"ベビーカー利用可否（stroller_ok）","example":true},"playground":{"type":"boolean","description":"遊具があるか（playground）","example":true},"athletics":{"type":"boolean","description":"アスレチックコースがあるか（athletics）","example":false},"waterPlay":{"type":"boolean","description":"水遊びが可能か（water_play）","example":true},"indoor":{"type":"boolean","description":"屋内施設か（indoor）","example":false}}},"ReviewListItemDto":{"type":"object","description":"レビュー一覧表示用のデータモデル","properties":{"id":{"type":"integer","format":"int64","description":"レビューID（reviews.id）","example":10},"spotId":{"type":"integer","format":"int64","description":"スポットID（reviews.spot_id）","example":3},"userId":{"type":"integer","format":"int64","description":"ユーザーID（reviews.user_id）","example":5},"userName":{"type":"string","description":"投稿者名（users.name）","example":"たろうパパ"},"childAgeGroup":{"type":"string","description":"子どもの年齢帯（単一選択）","example":"PRESCHOOL"},"rating":{"type":"integer","format":"int32","description":"総合評価（1〜5）","example":4},"ratingCost":{"type":"integer","format":"int32","description":"コスパ評価（1〜5）","example":4},"crowdLevel":{"type":"integer","format":"int32","description":"混雑度（1〜5）","example":2},"toiletCleanliness":{"type":"integer","format":"int32","description":"トイレ清潔度（1〜5）","example":5},"strollerEase":{"type":"integer","format":"int32","description":"ベビーカーの使いやすさ（1〜5）","example":4},"reviewText":{"type":"string","description":"レビュー本文","example":"敷地が広くて遊具の種類も多く、子どもが一日中遊べました"},"costTotal":{"type":"integer","format":"int32","description":"かかった合計金額（円）","example":1200},"createdAt":{"type":"string","format":"date-time","description":"投稿日時","example":"2025-12-01T10:30:00"},"updatedAt":{"type":"string","format":"date-time","description":"更新日時","example":"2025-12-01T12:00:00"}}},"SpotDetailDto":{"type":"object","description":"スポット詳細画面用のデータモデル","properties":{"id":{"type":"integer","format":"int64","description":"スポットID（spots.id）","example":2},"name":{"type":"string","description":"スポット名（spots.name）","example":"滋賀県希望ヶ丘文化公園"},"address":{"type":"string","description":"住所（spots.address）","example":"滋賀県野洲市北桜978"},"area":{"type":"string","description":"エリア名（spots.area）","example":"野洲市"},"priceType":{"type":"string","description":"予算区分（spots.price_type）","example":"1000円以内"},"categoryName":{"type":"string","description":"カテゴリ名（categories.name）","example":"公園"},"isFavorite":{"type":"boolean","description":"お気に入り済みかどうか（favorites）","example":true},"parkingInfo":{"type":"string","description":"駐車場情報（spots.parking_info）"},"toiletInfo":{"type":"string","description":"トイレ情報（spots.toilet_info）"},"targetAge":{"type":"string","description":"対象年齢（spots.target_age）","example":"小学校低学年まで"},"stayingTime":{"type":"string","description":"滞在時間目安（spots.staying_time）","example":"半日"},"convenienceStore":{"type":"string","description":"コンビニ情報（spots.convenience_store）"},"restaurantInfo":{"type":"string","description":"飲食店情報（spots.restaurant_info）"},"googleMapUrl":{"type":"string","description":"GoogleマップURL（spots.google_map_url）"},"closedDays":{"type":"string","description":"定休日（spots.closed_days）"},"officialUrl":{"type":"string","description":"公式サイトURL（spots.official_url）"},"notes":{"type":"string","description":"備考メモ（spots.notes）"},"diaperChanging":{"type":"boolean","description":"オムツ替えスペースの有無（diaper_changing）","example":true},"strollerOk":{"type":"boolean","description":"ベビーカー利用可否（stroller_ok）","example":true},"playground":{"type":"boolean","description":"遊具があるか（playground）","example":true},"athletics":{"type":"boolean","description":"アスレチックコースがあるか（athletics）","example":false},"waterPlay":{"type":"boolean","description":"水遊びが可能か（water_play）","example":true},"indoor":{"type":"boolean","description":"屋内施設か（indoor）","example":false}}},"CategoryDto":{"type":"object","description":"カテゴリ情報を表すデータモデル","properties":{"id":{"type":"integer","format":"int64","description":"カテゴリID（categories.id）","example":1},"name":{"type":"string","description":"カテゴリ名（categories.name）","example":"公園"}}}}}}